requirements.txt
Flask==3.0.0
yfinance==0.2.66
pandas==2.2.3
gunicorn==21.2.0
python-dotenv==1.0.0
pytz==2024.2
supabase==2.7.4
APScheduler==3.10.4
